feature(logged-in-user-voting-guide): update - type for percentage in entity

+++ election-backend/src/main/java/nl/hva/election_backend/service/VotingGuideResultsService.java
@@ -87,7 +87,7 @@ public class VotingGuideResultsService {
                         userId,
                         resultDto.getPartyId(),
                         resultDto.getPartyName(),
-                        resultDto.getPercentage()))
+                        (long) resultDto.getPercentage()))
                 .toList();
 //        Save all result entities
         votingGuideResultsRepository.saveAll(listOfResultEntities);
